Fixes transparency issues on first composited draw
When using compositing and transparency, the first draw was only
showing the shadow borders on the window. To fix this the opacity has
to be set after the first show. Don't know why this is needed, but it's
probably a bug in the underlying toolkits (maybe GTK or GDK).

It still looks a little jerky the first time the window is drawn.