Refresh the apps on icon-theme-changed
This shouldn't be necessary. The icons change by themselves when the theme changes, ithout the need to refresh the list of apps.

That said, at login, we're observing what looks like a race condition, which sometimes results in category and sometimes apps icons not being displayed.

The possible cause for this is a late initialization of the icon theme, so this commit might fix it.

If after this commit, this issue is still visible, then this commit should be reverted.