Highlight previous search so typing overwrites it (#191)
* Highlight previous search so typing overwrites it

When set to remember the last search, on next opening the menu it would display the last search and put the cursor at the end of it. If you wanted to do a new search you would first have to delete your previous search. More convenient if you starting typing it overwrites the last search and only if you place the cursor with arrow keys or mouse it switches to insert / append text to the last search.

There was existing code that did in fact do that but it was in the wrong place. This patch puts it in the right place. Fixes #86.

* Remove code to set position at end

Because it doesn't do anything when you select all the text.