Remove support for custom colors
Since the move to GTK3 custom color support is extremely hacky and buggy:

- It relies on gtk_widget_override_color() which is deprecated
- Only one call to gtk_widget_override_color() is functional, successive
calls are ignored, leading to the menu not changing colors when the custom
colors are changed or enabled/disabled.
- RegenPlugins leads to a loss of color.
- Theme changes are broken due to overridden colors.

The feature is also very niche and quite unecessary. If a custom color is
better than the default one we should apply the change in the theme or
define a new theme which inherits ours and simply defines the menu
differently.

The way to do this, in line with GTK's design, is via CSS at theme level.